fixed unlinking

+++ src/arctic3d/modules/pdb.py
@@ -202,6 +202,7 @@ def validate_api_hit(
             validated_pdbs.append((pdb_f, hit))
         else:
             log.debug(f"{pdb_id} failed validation")
+            os.unlink(pdb_f)
     return validated_pdbs
 
 
@@ -241,8 +242,9 @@ def get_maxint_pdb(validated_pdbs, interface_residues):
         log.info(f"pdb {pdb_f} retains the most interfaces ({max_nint})")
         # unlink pdb files
         for curr_pdb, curr_hit in validated_pdbs:
-            if curr_pdb != pdb_f:
-                os.unlink(curr_pdb)
+            if os.path.exists(curr_pdb):
+                if curr_pdb != pdb_f:
+                    os.unlink(curr_pdb)
         return pdb_f, hit, filtered_interfaces
     else:
         return None, None, None
